django_project/requirements.txt
aenum==2.1.2
amqp==1.4.9
anyjson==0.3.3
asgiref==2.3.2
asn1crypto==0.24.0
astor==0.7.1
async-timeout==3.0.1
attrdict==2.0.
attrs==18.2.0
autobahn==19.2.1
Automat==0.7.0
automium==0.2.6
automium-web==0.1.1
autopep8==1.4.3
base58==1.0.3
billiard==3.3.0.23
bitcoin==1.1.42
blessings==1.7
blockchain==1.4.4
bpython==0.17.1
bumpversion==0.5.3
celery==4.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorlog==4.0.2
constantly==15.1.0
coverage==4.5.2
coveralls==1.5.1
coz-bytecode==0.5.1
cryptography==2.4.2
curtsies==0.3.0
cycler==0.10.0
cytoolz==0.9.0.1
daphne==2.2.5
dataclasses==0.6
Django==2.1.7
django-crispy-forms==1.7.2
django-email-users==1.1.0
djangorestframework==3.9.1
djangorestframework-stubs==0.2.0
docopt==0.6.2
ecdsa==0.13
enum-compat==0.0.2
eth-abi==2.0.0b6
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.1
eth-rlp==0.1.2
eth-typing==2.0.0
eth-utils==1.4.1
ethpm==0.1.4a12
Events==0.3
Flask==1.0.2
furl==2.0.0
future==0.17.1
gevent==1.4.0
greenlet==0.4.15
hexbytes==0.1.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipfsapi==0.4.3
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
klein==17.10.0
legacy==0.1.6
logzero==1.5.0
lru-dict==1.1.6
MarkupSafe==1.1.0
memory-profiler==0.55.0
mmh3==2.5.1
mock==2.0.0
mpmath==1.1.0
neo-boa==0.5.6
neo-python==0.8.4
neo-python-rpc==0.2.1
neocore==0.5.6
numpy==1.16.2
orderedmultidict==1.0
parsimonious==0.8.1
pbr==5.1.1
peewee==3.8.1
pexpect==4.6.0
Pillow==5.4.1
pluggy==0.8.1
plyvel==1.0.5
prompt-toolkit==2.0.7
protobuf==3.7.0
psutil==5.4.8
psycopg2-binary==2.7.7
ptyprocess==0.6.0
py==1.7.0
pycodestyle==2.4.0
pycparser==2.19
pycryptodome==3.7.2
Pygments==2.3.1
PyHamcrest==1.9.0
pymitter==0.2.3
Pympler==0.6
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
redis==2.10.3
requests==2.21.0
rlp==1.1.0
scipy==1.2.1
scrypt==0.8.6
six==1.12.0
toolz==0.9.0
tqdm==4.29.1
Twisted==18.9.0
txaio==18.8.1
typing==3.6.6
urllib3==1.24.1
views==0.3
virtualenv==16.2.0
wcwidth==0.1.7
web3==5.0.0a6
websockets==7.0
Werkzeug==0.14.1
zope.interface==4.6.0
